[FEATURE] Adding tool prepare_sam in app/rabema.
This tool allows to fix SAM records for Rabema.  The SAM standard recommends not
to store the sequence and quality for secondary alignments to reduce the file
size.  The new RazerS version creates such output and since we do not want to
use the Fragment Store in the future (which gets rid of this problem) but plain
BAM I/O, we need to expand the records.

A future version of this tool should process both SAM and BAM because all
SAM/BAM sorting tools work on BAM files only.


git-svn-id: http://svn.seqan.de/seqan/trunk@11784 e6417c60-b987-48fd-844e-b20f0fcc1017