[FEATURE] `convertMatchesToGlobal()` now fills qualities `alignQualityStore`.
The function `convertMatchesToGlobal()` already computed alignments to it was natural to extend it to also fill qualities in the `alignQualityStore` member of `FragmentStore`.
The `errors` member is set as the number of edit distance errors, the `score` member is set as the edit distance alignment score.
For the case of giving scaled scores (e.g. 1000 for matches, 1001 for indels so matches are slightly preferred over indels), we normalize by the mismatch score.


git-svn-id: http://svn.seqan.de/seqan/trunk@11836 e6417c60-b987-48fd-844e-b20f0fcc1017